refactor: consolidate OAuth and universal link views into Login module
Move OAuth and universal link related views, hooks, and tests from src/app/auth to src/views/Login to centralize authentication UI components. Remove duplicate UniversalLinkView component and update all import paths to use the consolidated services structure.
