+++ src/views/Login/OAuthLinkView.tsx
@@ -1,6 +1,6 @@
-import { UserSettings } from '@internxt/sdk/dist/shared/types/userSettings';
 import { Button } from '@internxt/ui';
 import authService from 'services/auth.service';
+import oauthService from 'services/oauth.service';
 import localStorageService from 'app/core/services/local-storage.service';
 import navigationService from 'app/core/services/navigation.service';
 import { AppView } from 'app/core/types';
@@ -8,14 +8,11 @@ import { useTranslationContext } from 'app/i18n/provider/TranslationProvider';
 import InternxtLogo from 'assets/icons/big-logo.svg?react';
 import { useEffect, useMemo } from 'react';
 
-const DEEPLINK_SUCCESS_REDIRECT_BASE = 'internxt://login-success';
-
-export default function UniversalLinkView(): JSX.Element {
+const OAuthLinkView = (): JSX.Element => {
   const { translate } = useTranslationContext();
   const user = useMemo(() => localStorageService.getUser(), []);
 
-  const urlParams = new URLSearchParams(window.location.search);
-  const redirectUri = urlParams.get('redirectUri');
+  const urlParams = new URLSearchParams(globalThis.location.search);
 
   useEffect(() => {
     if (!user) {
@@ -24,31 +21,27 @@ export default function UniversalLinkView(): JSX.Element {
     }
   }, [user]);
 
-  const getUniversalLinkAuthUrl = (user: UserSettings) => {
-    const token = localStorageService.get('xToken');
-    const newToken = localStorageService.get('xNewToken');
-    if (!token) return AppView.Login;
-    if (!newToken) return AppView.Login;
-
-    let baseURL = DEEPLINK_SUCCESS_REDIRECT_BASE;
-    if (redirectUri) {
-      baseURL = Buffer.from(redirectUri, 'base64').toString();
-    }
-
-    return `${baseURL}?mnemonic=${btoa(user.mnemonic)}&token=${btoa(token)}&newToken=${btoa(
-      newToken,
-    )}&privateKey=${btoa(user.privateKey)}`;
-  };
-
-  // Should redirect to login in the useEffect
   if (!user) return <></>;
 
   const handleGoToLogin = () => {
-    authService.logOut();
+    const params: Record<string, string> = {};
+    urlParams.forEach((value, key) => {
+      params[key] = value;
+    });
+    authService.logOut(params);
   };
 
-  const handleGoToUniversalLinkUrl = () => {
-    window.location.href = getUniversalLinkAuthUrl(user);
+  const handleContinueWithCurrentUser = () => {
+    const newToken = localStorageService.get('xNewToken');
+    if (!newToken) {
+      navigationService.history.replace(AppView.Login);
+      return;
+    }
+
+    const success = oauthService.sendAuthSuccess(user, newToken);
+    if (!success) {
+      navigationService.history.replace(AppView.Login);
+    }
   };
 
   return (
@@ -61,22 +54,21 @@ export default function UniversalLinkView(): JSX.Element {
           <div className="mb-6 flex justify-center">
             <InternxtLogo className="h-auto w-52 text-gray-100" />
           </div>
-          <h2 className="text-center text-xl font-medium text-gray-100">{translate('auth.universalLink.loginAs')}</h2>
+          <h2 className="text-center text-xl font-medium text-gray-100">{translate('auth.oauthLink.continueAs')}</h2>
           <h3
             title={user.email}
             className="over mb-6 overflow-x-hidden text-ellipsis text-center text-xl font-medium text-gray-60"
           >
             {user.email}
           </h3>
-          {/* Universal links needs to be clicked in order to work, JS window.open does not work */}
-          <Button onClick={handleGoToUniversalLinkUrl} className="w-full">
-            {translate('auth.universalLink.openApp')}
+          <Button onClick={handleContinueWithCurrentUser} className="w-full">
+            {translate('auth.oauthLink.continue')}
           </Button>
           <div className="separator my-6"></div>
           <div className="flex flex-row justify-center">
-            <h4 className="text-base font-medium">{translate('auth.universalLink.anotherAccount')}</h4>
+            <h4 className="text-base font-medium">{translate('auth.oauthLink.anotherAccount')}</h4>
             <button onClick={handleGoToLogin} className="ml-2.5 text-base font-medium no-underline">
-              {translate('auth.universalLink.login')}
+              {translate('auth.oauthLink.login')}
             </button>
           </div>
         </div>
@@ -99,4 +91,6 @@ export default function UniversalLinkView(): JSX.Element {
       </div>
     </main>
   );
-}
+};
+
+export default OAuthLinkView;
