+++ src/views/Login/hooks/useOAuthFlow.ts
@@ -3,7 +3,7 @@ import localStorageService from 'app/core/services/local-storage.service';
 import navigationService from 'app/core/services/navigation.service';
 import { AppView } from 'app/core/types';
 import { useEffect } from 'react';
-import oauthService from '../services/oauth.service';
+import oauthService from 'services/oauth.service';
 
 interface UseOAuthFlowParams {
   authOrigin: string | null;
