+++ src/views/Login/index.ts
@@ -3,3 +3,6 @@ export { default as UniversalLinkSuccessView } from './UniversalLinkSuccessView'
 export { default as BlockedAccountView } from './BlockedAccountView';
 export { default as RecoveryLinkView } from './RecoveryLinkView';
 export { default as RecoverAccountView } from './RecoverAccountView';
+export { default as OAuthLinkView } from './OAuthLinkView';
+export { default as UniversalLinkErrorView } from './UniversalLinkErrorView';
+export { default as UniversalLinkOkView } from './UniversalLinkOkView';
